feat: make the landign page ssr
