+++ apps/website/components/page-components/landing-page/Inspiration/index.tsx
@@ -1,4 +1,3 @@
-'use client';
 import { Box } from '@/components/ui/box';
 import { Heading } from '@/components/ui/heading';
 import { Text } from '@/components/ui/text';
